Fix for maps with null keys.
Right now this code assumes that maps would not have null keys. However, dart allows maps to have null keys. Right now, the value corresponding to the null key is used as the next key, and everything in the list is shifted left by one. More explicitly enforcing odd and even entries with a bool avoids this bug.